# packages/hook-contract/openapi/sink.yaml
openapi: 3.1.0
info:
  title: OrchestKit Hook Sink API
  version: 0.1.1
  summary: HTTP ingest endpoint for HMAC-signed OrchestKit hook events.
  description: >-
    Wire contract for an HTTP sink that receives OrchestKit hook deliveries.


    Each request carries a Claude Code hook event envelope signed with HMAC-SHA256 over the **raw
    request bytes**. The signature travels in the `X-CC-Hooks-Signature` header using the scheme
    `t=<unix-seconds>,v1=<hex-hmac>` (comma-separated; multiple `vN=` schemes may appear during key
    rotation). The canonical event JSON Schemas and the reference signer/verifier ship in the
    `@orchestkit/hook-contract` npm/PyPI package; a runnable Express example lives in
    `examples/generic-client/`.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: OrchestKit
    url: https://github.com/yonatangross/orchestkit
servers:
  - url: https://sink.orchestkit.dev
    description: Your hook sink deployment (replace with your own host).
tags:
  - name: hooks
    description: Hook event ingestion.
security:
  - HookSignature: []
paths:
  /hooks/ingest:
    post:
      operationId: ingestHookEvent
      tags:
        - hooks
      summary: Ingest one signed hook event.
      description: >-
        Verify the `X-CC-Hooks-Signature` header against the raw body, then accept the event for
        asynchronous processing. The HMAC is computed over the exact bytes received, so the body
        MUST NOT be re-serialized in transit.
      requestBody:
        required: true
        description: A single hook event envelope (see `HookEvent`).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HookEvent'
      responses:
        '202':
          description: Accepted — the event was verified and queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accepted'
        '400':
          description: Bad request — missing/malformed signature header or non-JSON body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '401':
          description: Unauthorized — stale timestamp or HMAC signature mismatch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '429':
          description: Too many requests — the sink is rate limiting.
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
                minimum: 0
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
components:
  securitySchemes:
    HookSignature:
      type: apiKey
      in: header
      name: X-CC-Hooks-Signature
      description: >-
        HMAC-SHA256 signature of the raw request body. Format: `t=<unix-seconds>,v1=<hex-hmac>`. The
        signed payload is `"<t>." + rawBody`. Multiple `vN=` schemes may be present during key
        rotation; any valid one authenticates the request. Verifiers reject timestamps outside a
        replay window (default 300s).
  schemas:
    SessionStartEvent:
      title: SessionStart hook event
      type: object
      properties:
        event:
          type: string
          const: SessionStart
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            plugin_root:
              type: string
              description: Absolute path to the plugin root (injected by run-hook.mjs)
            permissionMode:
              type: string
              description: Permission mode (default | acceptEdits | dontAsk | auto)
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    UserPromptSubmitEvent:
      title: UserPromptSubmit hook event
      type: object
      properties:
        event:
          type: string
          const: UserPromptSubmit
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            prompt:
              type: string
              description: The user prompt text being submitted
          additionalProperties: true
          required:
            - prompt
      required:
        - event
      additionalProperties: true
    PreToolUseEvent:
      title: PreToolUse hook event
      type: object
      properties:
        event:
          type: string
          const: PreToolUse
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            tool_name:
              type: string
              description: The tool being invoked (e.g. Bash, Write, Read)
            tool_input:
              type: object
              additionalProperties: true
              description: Tool-specific input parameters
            tool_use_id:
              type: string
              description: CC 2.1.69 correlation ID for matching with PostToolUse
          additionalProperties: true
          required:
            - tool_name
            - tool_input
      required:
        - event
      additionalProperties: true
    PostToolUseEvent:
      title: PostToolUse hook event
      type: object
      properties:
        event:
          type: string
          const: PostToolUse
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            tool_name:
              type: string
              description: The tool that was invoked
            tool_input:
              type: object
              additionalProperties: true
              description: Tool input passed to the invocation
            tool_output:
              description: Tool output (type varies by tool)
            tool_error:
              type: string
              description: Tool error message if one was produced
            exit_code:
              type: number
              description: Tool exit code where applicable
            tool_use_id:
              type: string
              description: CC 2.1.69 correlation ID
            duration_ms:
              type: number
              description: Tool execution duration in milliseconds
          additionalProperties: true
          required:
            - tool_name
            - tool_input
      required:
        - event
      additionalProperties: true
    PostToolUseFailureEvent:
      title: PostToolUseFailure hook event
      type: object
      properties:
        event:
          type: string
          const: PostToolUseFailure
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            tool_name:
              type: string
              description: The tool that failed
            tool_input:
              type: object
              additionalProperties: true
              description: Tool input passed to the failed invocation
            tool_error:
              type: string
              description: Error message from the tool
            exit_code:
              type: number
              description: Tool exit code
            tool_use_id:
              type: string
              description: CC 2.1.69 correlation ID
          additionalProperties: true
          required:
            - tool_name
            - tool_input
            - tool_error
      required:
        - event
      additionalProperties: true
    PermissionRequestEvent:
      title: PermissionRequest hook event
      type: object
      properties:
        event:
          type: string
          const: PermissionRequest
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            tool_name:
              type: string
              description: The tool requesting permission
            tool_input:
              type: object
              additionalProperties: true
              description: Tool input that triggered the permission request
            permission_suggestions:
              type: object
              additionalProperties: true
              description: CC 2.1.69 suggested "always allow" options
          additionalProperties: true
          required:
            - tool_name
            - tool_input
      required:
        - event
      additionalProperties: true
    SubagentStartEvent:
      title: SubagentStart hook event
      type: object
      properties:
        event:
          type: string
          const: SubagentStart
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            subagent_type:
              type: string
              description: Subagent type (CC 2.1.0+ field name)
            agent_type:
              type: string
              description: Alternative agent type field name
            agent_id:
              type: string
              description: Agent ID
            parent_agent_id:
              type: string
              description: Parent agent ID for trace stitching (CC 2.1.139)
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    SubagentStopEvent:
      title: SubagentStop hook event
      type: object
      properties:
        event:
          type: string
          const: SubagentStop
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            agent_type:
              type: string
              description: Agent type
            agent_id:
              type: string
              description: Agent ID
            agent_output:
              type: string
              description: Final agent output text
            output:
              type: string
              description: Alternative output field name
            error:
              type: string
              description: Error from subagent if it failed
            duration_ms:
              type: number
              description: Total subagent execution duration
            agent_transcript_path:
              type: string
              description: Path to subagent transcript (CC 2.1.69)
            last_assistant_message:
              type: string
              description: Final assistant message text (CC 2.1.47+)
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    StopEvent:
      title: Stop hook event
      type: object
      properties:
        event:
          type: string
          const: Stop
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            last_assistant_message:
              type: string
              description: Final assistant message text (CC 2.1.47+)
            stop_hook_active:
              type: boolean
              description: Whether a stop hook is currently active (prevents re-entry)
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    SessionEndEvent:
      title: SessionEnd hook event
      type: object
      properties:
        event:
          type: string
          const: SessionEnd
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    SetupEvent:
      title: Setup hook event
      type: object
      properties:
        event:
          type: string
          const: Setup
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    NotificationEvent:
      title: Notification hook event
      type: object
      properties:
        event:
          type: string
          const: Notification
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            message:
              type: string
              description: Notification message text
            notification_type:
              type: string
              description: Notification type discriminator
          additionalProperties: true
          required:
            - message
      required:
        - event
      additionalProperties: true
    PreCompactEvent:
      title: PreCompact hook event
      type: object
      properties:
        event:
          type: string
          const: PreCompact
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            compaction_count:
              type: number
              description: Number of compactions in this session so far
            context_size_after:
              type: number
              description: Estimated context size after compaction (tokens)
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    TeammateIdleEvent:
      title: TeammateIdle hook event
      type: object
      properties:
        event:
          type: string
          const: TeammateIdle
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            teammate_id:
              type: string
              description: Teammate agent ID
            teammate_type:
              type: string
              description: Teammate agent type
            idle_duration_ms:
              type: number
              description: How long the teammate has been idle
          additionalProperties: true
          required:
            - teammate_id
            - teammate_type
            - idle_duration_ms
      required:
        - event
      additionalProperties: true
    TaskCompletedEvent:
      title: TaskCompleted hook event
      type: object
      properties:
        event:
          type: string
          const: TaskCompleted
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          properties:
            task_id:
              type: string
              description: Completed task ID
            task_subject:
              type: string
              description: Task subject
            task_status:
              type: string
              description: Task result status
            token_count:
              type: number
              description: Tokens consumed by the task (CC 2.1.30)
            tool_uses:
              type: number
              description: Tool invocations in the task (CC 2.1.30)
          additionalProperties: true
          required:
            - task_id
      required:
        - event
      additionalProperties: true
    InstructionsLoadedEvent:
      title: InstructionsLoaded hook event
      type: object
      properties:
        event:
          type: string
          const: InstructionsLoaded
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    WorktreeCreateEvent:
      title: WorktreeCreate hook event
      type: object
      properties:
        event:
          type: string
          const: WorktreeCreate
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    WorktreeRemoveEvent:
      title: WorktreeRemove hook event
      type: object
      properties:
        event:
          type: string
          const: WorktreeRemove
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    ConfigChangeEvent:
      title: ConfigChange hook event
      type: object
      properties:
        event:
          type: string
          const: ConfigChange
        timestamp:
          type: string
          description: ISO-8601 UTC timestamp
        session_id:
          type: string
        cwd:
          type: string
        payload:
          type: object
          additionalProperties: true
      required:
        - event
      additionalProperties: true
    HookEvent:
      title: HookEvent
      description: >-
        Any OrchestKit hook event envelope. The canonical JSON Schemas live in
        @orchestkit/hook-contract (`HOOK_EVENT_SCHEMAS`); these mirror them.
      oneOf:
        - $ref: '#/components/schemas/SessionStartEvent'
        - $ref: '#/components/schemas/UserPromptSubmitEvent'
        - $ref: '#/components/schemas/PreToolUseEvent'
        - $ref: '#/components/schemas/PostToolUseEvent'
        - $ref: '#/components/schemas/PostToolUseFailureEvent'
        - $ref: '#/components/schemas/PermissionRequestEvent'
        - $ref: '#/components/schemas/SubagentStartEvent'
        - $ref: '#/components/schemas/SubagentStopEvent'
        - $ref: '#/components/schemas/StopEvent'
        - $ref: '#/components/schemas/SessionEndEvent'
        - $ref: '#/components/schemas/SetupEvent'
        - $ref: '#/components/schemas/NotificationEvent'
        - $ref: '#/components/schemas/PreCompactEvent'
        - $ref: '#/components/schemas/TeammateIdleEvent'
        - $ref: '#/components/schemas/TaskCompletedEvent'
        - $ref: '#/components/schemas/InstructionsLoadedEvent'
        - $ref: '#/components/schemas/WorktreeCreateEvent'
        - $ref: '#/components/schemas/WorktreeRemoveEvent'
        - $ref: '#/components/schemas/ConfigChangeEvent'
      discriminator:
        propertyName: event
        mapping:
          SessionStart: '#/components/schemas/SessionStartEvent'
          UserPromptSubmit: '#/components/schemas/UserPromptSubmitEvent'
          PreToolUse: '#/components/schemas/PreToolUseEvent'
          PostToolUse: '#/components/schemas/PostToolUseEvent'
          PostToolUseFailure: '#/components/schemas/PostToolUseFailureEvent'
          PermissionRequest: '#/components/schemas/PermissionRequestEvent'
          SubagentStart: '#/components/schemas/SubagentStartEvent'
          SubagentStop: '#/components/schemas/SubagentStopEvent'
          Stop: '#/components/schemas/StopEvent'
          SessionEnd: '#/components/schemas/SessionEndEvent'
          Setup: '#/components/schemas/SetupEvent'
          Notification: '#/components/schemas/NotificationEvent'
          PreCompact: '#/components/schemas/PreCompactEvent'
          TeammateIdle: '#/components/schemas/TeammateIdleEvent'
          TaskCompleted: '#/components/schemas/TaskCompletedEvent'
          InstructionsLoaded: '#/components/schemas/InstructionsLoadedEvent'
          WorktreeCreate: '#/components/schemas/WorktreeCreateEvent'
          WorktreeRemove: '#/components/schemas/WorktreeRemoveEvent'
          ConfigChange: '#/components/schemas/ConfigChangeEvent'
    Accepted:
      title: Accepted
      description: Event accepted and queued for processing.
      type: object
      required:
        - ok
        - event
      additionalProperties: false
      properties:
        ok:
          type: boolean
          enum:
            - true
        event:
          type: string
          description: Echoed `event` field of the accepted envelope.
    BadRequestError:
      title: BadRequestError
      description: Signature header missing/malformed, or the verified body was not valid JSON.
      type: object
      required:
        - ok
        - reason
      additionalProperties: false
      properties:
        ok:
          type: boolean
          enum:
            - false
        reason:
          type: string
          enum:
            - missing_header
            - malformed_header
            - invalid_json
    UnauthorizedError:
      title: UnauthorizedError
      description: 'Signature verification failed: timestamp outside the replay window, or HMAC mismatch.'
      type: object
      required:
        - ok
        - reason
      additionalProperties: false
      properties:
        ok:
          type: boolean
          enum:
            - false
        reason:
          type: string
          enum:
            - stale
            - signature_mismatch
    RateLimitError:
      title: RateLimitError
      description: The sink is shedding load; retry after the indicated delay.
      type: object
      required:
        - ok
        - reason
      additionalProperties: false
      properties:
        ok:
          type: boolean
          enum:
            - false
        reason:
          type: string
          enum:
            - rate_limited